+++ types/database.types.ts
@@ -1494,7 +1494,13 @@ export type Database = {
     }
     Enums: {
       activity_target: "caang" | "anggota"
-      attendance_status: "hadir" | "izin" | "sakit" | "alfa" | "telat"
+      attendance_status:
+        | "hadir"
+        | "izin"
+        | "sakit"
+        | "alfa"
+        | "telat"
+        | "magang"
       gender_type: "L" | "P"
       piket_day:
         | "Senin"
@@ -1646,7 +1652,7 @@ export const Constants = {
   public: {
     Enums: {
       activity_target: ["caang", "anggota"],
-      attendance_status: ["hadir", "izin", "sakit", "alfa", "telat"],
+      attendance_status: ["hadir", "izin", "sakit", "alfa", "telat", "magang"],
       gender_type: ["L", "P"],
       piket_day: [
         "Senin",